requirements.txt
fastapi
uvicorn
torch
torchvision
timm
pillow
numpy
opencv-python
streamlit
requests
python-multipart
